remove validate schema & some mocks
