--- __mocks__/creator/validate-options.mock.js
@@ -1,23 +1,0 @@
-"use strict";
-
-const fs = require("fs");
-const path = require("path");
-
-function getPath(part) {
-	return path.join(process.cwd(), part);
-}
-
-function validateOptions(opts) {
-	return Object.keys(opts).forEach(location => {
-		let part = getPath(opts[location]);
-		try {
-			fs.readFileSync(part);
-		} catch (err) {
-			throw new Error("Did not find the file");
-		}
-	});
-}
-
-module.exports = {
-	validateOptions
-};
