+++ lib/commands/migrate.js
@@ -6,9 +6,6 @@ const diff = require("diff");
 const inquirer = require("inquirer");
 const PLazy = require("p-lazy");
 const Listr = require("listr");
-const validateSchema = require("../utils/validateSchema.js");
-const webpackOptionsSchema = require("../utils/webpackOptionsSchema");
-const WebpackOptionsValidationError = require("../utils/WebpackOptionsValidationError");
 
 module.exports = function transformFile(
 	currentConfigPath,
@@ -82,17 +79,8 @@ module.exports = function transformFile(
 				.then(answers => {
 					if (answers["confirmMigration"]) {
 						fs.writeFile(outputConfigPath, result, "utf8", err => {
-							const webpackOptionsValidationErrors = validateSchema(
-								webpackOptionsSchema,
-								require(outputConfigPath)
-							);
 							if (err) {
 								throw err;
-							} else if (webpackOptionsValidationErrors.length) {
-								const validationMsg = new WebpackOptionsValidationError(
-									webpackOptionsValidationErrors
-								);
-								throw validationMsg.message;
 							} else {
 								console.log(
 									chalk.green(
